About it
This is a Frankenapp made of bits and pieces I found here and there around WebView that I made in order to interact with  pixelfed.social instead of the browser on the phone. One [accidentally]added feature is the possibility of actually taking a picture and uploading it besides just selecting one one from the phone.
I'm not a programmer.